[UsdImaging] Remove calls to MarkAllCollectionsDirty.
MarkAllCollectionsDirty was a big hammer mechanism to cause Dirty List and Batches to be re-evaluated.
Hydra often relied on the scene delegate to know these we needed, but over time the logic has move to Hydra's core and become more fine grained where possible.

This change purges the last remaining calls to this function for the UsdImaging Delegate.

There was 3 cases: SetReprFallback, SetCullStyleFallback and SetDisplayGuides.

For SetReprFallback, the logic to invalidate collections in now in the change tracker when a rprim is marked dirty, so the call is redundant.
For SetCullStyleFallback, previous changes made sure the batches were marked dirty if the geometry shader is changed. So the call is not needed.

SetDisplayGuides is the most tricky.  This call was just marking collections dirty.  However, really what it is doing is changing some prims render tags.
The code was never marking the prims render tag dirty, which would have broke any backend that cached the render tag.
So this change, adds a new API on the adapters to mark the render tag dirty.  The change tracker already handles propagation to cause dirty list and batch re-evaluation
if a prims render tag is dirty.  Therefore the last MarkAllCollectionsDirty can be removed!

Changes:
- Removed calls to MarkAllCollectionsDirty from the UsdImaging Delegate
- SetDisplayGuides now calls MarkRenderTagDirty for each prim.
- Added MarkRenderTagDirty to UsdImagingPrimAdapter.  Default implementation is a no-op.
- GprimAdapters implementation of MarkRenderTagDirty dirties the tag in the render index.
- Instancer and skinning implementation forward the message to contained adapters

(Internal change: 1957600)