[Hd/HdSt/Hdx] Remove MarkAllCollectionsDirty.
MarkAllCollectionsDirty was a big hammer mechanism to cause Dirty List and Batches to be re-evaluated.

This is a final step to remove this API.

RenderIndex::Clear() was calling it, but the change tracker prim removal calls did the same work, so it was redundant.

Several tests were still using it:
- The HideRprim/UnhideRprim in the Hd Unit test delegate was using it, but should of just called MarkRprimDirty with the dirty render tag.
- testHdxPickingWithReprAndRefineChanges, was XXX, because the change tracker wasn't handling repr differences correctly.  Now the change tracker is fixed, these calls can be removed.

Changes:
- Removed MarkAllCollectionsDirty from change tracker
- Removed call to MarkAllCollectionsDirty from HdRenderIndex::Clear()
- Replaced calls to MarkAllCollectionsDirty with MarkRprimDirty in HdUnitTestDelegate::(Un)HideRprim()
- Removed calls to MarkAllCollectionsDirty from testHdxPickingWithReprAndRefineChanges

(Internal change: 1957606)