Move functionality from SdfSchema to SdfSchemaBase
This separates the singleton instance SdfSchema from
the SdfSchemaBase class that contains all the
functionality. This makes it easier to derive a new
schema that extends the standard fields and types
defined in Sdf.

As part of this change, the parsing of plugin metadata
values and the initialization of the SdfValueTypeNames
struct were reworked to avoid reentrancy into SdfSchema.
This previously required calling SetInstancedConstructed
to avoid deadlocks, which in theory this could cause
concurrency issues if multiple threads asked the schema
for field information before the schema was fully
initialized.

(Internal change: 1957725)