[UsdSkelImaging] Fix resync and removal for skinned prims/skeletons
Previously, a skinned prim resync used the default resync handling in PrimAdapter, which removes/readds the skinned prim. This broke the hijacking setup used to route callbacks for skinned prims via the SkeletonAdapter.

- Add handling for Resync in SkeletonAdapter:
Resyncing a skinned prim removes the skinned prim and its computations, and repopulates the skel root, which in turn sets up the hijacking behavior necessary while populating the skinned prim. The computations are added when populating the Skeleton.
While this may seem like a big hammer, only prims that were removed are re-inserted into the Hydra render index. The advantage here is that UsdImaging is oblivious to our hijacking games.
Resyncing a skeleton removes the bone mesh and all the skinned prims targeted by the skeleton (and their computations).

- Update test with scenarios to trigger resync for the skinned prim and the skeleton.

(Internal change: 1957996)