[UsdImagingGL] Ensure consistent Render Tag usage in UsdImaging's Engine.
Hydra currently has multiple paths in which render tags can be provided.  Obviously this is risky for consistency reasons.

Having render tags being part of the collection has also been a contributing factor to the "Collection Paradox" issue, where Hydra doesn't necessarily have all the collections available in order to work out what to sync.

This change moves the code that was computing the render tag into its own function (rather than embedded in _UpdateHydraCollection).

This function is then used to set the render tags on the collection and for the different task params.  It is also being used to the set the render tags on the collection, while the dataflows are transiting.

Changes:
- UsdImagingGLEngine no longer stores a persistent _renderTags member,  the render tags is built in place as needed.
- _UpdateHydraCollection() no longer takes a pointer to the _renderTags member.
- Added new function _ComputeRenderTags to fill in render tags based on the render params.
- UsdImagingGLEngine::_Render() no longer sets the HdxTokens->renderTags task context data.  This key is meant for internal communication of render tags from the render setup task to the render task.

(Internal change: 1958258)