[Hd] Add Render Tag Version Tracking.
This change is part of resolving the collection paradox.

In order to break the paradox, the sync system will no longer be working directly with collections.  So for dirty list management, we need to track when render tags change.
This versioning combined with the prim index versions will indicate when the dirty list needs to be rebuilt.

Management of render tags is also moving into tasks, so the change to render pass is transitional.

Changes:
- Added _renderTagVersion to change tracker, with MarkRenderTagsDirty() to increment it and GetRenderTagVersion to obtain the current version.
- If an rprim's render tag changes in the change tracker.  Increment the render tag version number.
- Render Pass currently increments the render tag version if the render tags in the collection have changed.

(Internal change: 1958395)