[pxrUsdMayaGL,usdMaya] add batch renderer API for getting/setting the selection resolution
This also adds an attribute on the pxrHdImagingShape that can be used to
change the selection resolution between a few pre-defined options (256x256,
512x512, 1024x1024, 2048x2048, or 4096x4096). Previously, the resolution
was hard-coded at 256x256 and artists were noticing issues with very small
objects being missed in marquee selections. This new control should make
it easier to do some experimentation with different values as we try to
dial in some better settings for selections.

A PXRMAYAHD_DEFAULT_SELECTION_RESOLUTION TfEnvSetting was added for
choosing one of these pre-defined options as the default.

(Internal change: 1958448)