[usdMaya] avoid unnecessary authoring and better behavior with non-writable nodes during shader export
At least currently, many prim writers (including shader writers) define
their prims and do some amount of authoring in their constructors, so
merely creating the prim writer can have the unfortunate side effect of
creating scene description unnecessarily, even if Write() is never called
on it.

While that should be addressed separately, we can also take advantage of
the direction of the dependency graph traversal and discard some plugs
before we consider authoring shaders for them. The traversal visits shading
nodes from destinations to sources, so if we encounter a plug for which we
do not have a source, we do not need to consider any of that plug's
destinations.

This also ensures that we ignore any nodes that are not writable. In
particular, default nodes like Maya's "defaultColorMgtGlobals" often have
connections to many more shaders than just what's in the current
shadingEngine's dependency graph. While traversing the shadingEngine's
dependency graph, we may end up visiting plugs that connect
"defaultColorMgtGlobals" as a source to some un-related shader as a
destination. The changes here avoid polluting the namespace hierarchy of
the current Material prim being authored with irrelevant shader prims.

(Internal change: 1958652)