[Hd] Simplify Dirty List's transition management.
As part of the "Collection Paradox" series of changes, this change simplifies Dirty List's transitions..

Prior to this change, GetDirtyRprims() could be called multiple times during sync() and the dirty list was expected to return the same dirty list as it did when it was built.
By removing these additional callers in previous change list, GetDirtyRprims() is only called once.  Which means, the dirty list no longer needs an explicit Clear call to move it to
the Empty state.

With this change, the Dirty List always starts and ends in the Empty state and one of 4 actions can be performed based on which inputs have changed.
- An empty list, if nothing has changed
- Everything that is dirty if the filter has changed
- A new varying list if the varying set has changed
- The cached varying list otherwise.

The definition of the "filter changes" has also been modified by the change.  Previously the dirty list worked on collection version changes.
Now it works on Rprim Index and render tag versions.

Changes:
- Removed HdDirtyList::Clear()
- Renamed _UpdateIDs() to _BuildDirtyList() and removed id's param (the function uses the _dirtyIds member directly)
- Replaced _collectionVersion with _rprimIndexVersion and _renderTagVersion.
- Updated doxygen comment for HdDirtyList class as it was getting stale.
- Removed _isEmpty boolean.
- Updated GetDirtyRprims() description
- Simplified GetDirtyRprims early out conditions as it no longer needs to distinguish between multiple calls in the same sync.
- The Init dirty list build step, can now update the version tracking rather than waiting until Clear().
- Removed collection name from Debug messages.
- Removed pointless no-op else clause for scene state is not equal (that had to always be true and didn't do anything.
- Removed Collection path dump for debug output as dirty list no longer uses this.
- Render Tag version in the change tracker wasn't being initialized.
- Several Tests needed updating because
    - Clear function removal
    - They were using "Sync" to clean dirty bits, but that doesn't work after getting the dirty list (if your getting the list, you will clean the items)

(Internal change: 1958918)