[pxrUsdMayaGL,usdMaya] add batch renderer API for enabling/disabling depth selection
Hydra does not currently support selection of occluded objects natively, so
it is up to the client to implement a solution. For area/marquee selections
in Maya, the current implementation comes with a significant performance
penalty if the number of objects grows large, so it is disabled by default.
The PXRMAYAHD_ENABLE_DEPTH_SELECTION env setting has been available for a
while and can be used to turn depth selection on, but this was previously
the only way to enable it, and once enabled, it could not be changed in a
Maya session.

This change moves the env setting onto the pxrHdImagingShape and adds an
attribute on the shape that will toggle depth selection on or off. The env
setting now simply sets the default value for this attribute.

(Internal change: 1958958)