[Hd] Pass Render Tags as input to dirty list.
This change is part of a series of changes in to fix the "collection paradox".

This change removes dirty list's dependence on having a collection, by passing in render tags as an input parameter rather than getting them from a stored container.

The collection member of dirty list can not be completely removed at this point, because external systems are still querying for collection from the dirty list.

Changes:
- GetDirtyRprims() now takes the render tags as a parameter.
- The render tags as flowed into _BuildDirtyList() and stored in the filter param instead of the collection.
- In _DirtyListFilterPredicate() rather than using the collection, the predicate searches the render tag list itself.
- Render Index is currently fetching the render tags for dirty list from the collections.  This will be addressed later.

(Internal change: 1959127)