[Hd] Remove GetCollection() API from HdDirtyList.
This change is part of a series of changes in to fix the "collection paradox".

This change removes the ability for external systems to get access to HdDirtyList's collection.
With other changes in the series focusing on removing HdDirtyList own dependence on the this collection, this gets closer to the goal of removing collection from HdDirtyList.

In order to facilitate the change, RenderIndex sync queue now stores the dirty list and collection.

Changes:
- Removed GetCollection() from HdDirtyList.
- HdRenderIndex::Sync() now takes both a collection and dirty list.
- Renamed _DirtyListVector to _SyncQueue.  It now contains both a dirty list and collection in each entry.
- HdRenderIndex's SyncAll() now operates on the collection from the sync queue rather than obtaining it from the dirty list.
- Updated HdRenderPass::Sync() to pass the collection to the render index.

(Internal change: 1959272)