Prefer authored file format targets during composition
Previously, if an asset path for a composition arc like
references, payloads, or sublayers contained a "target"
argument, e.g.:

"references = @foo.usd:SDF_FORMAT_ARGS:target=foo@"

it would be ignored in favor of the default target
specified to the PcpCache constructor. The authored
target now takes precedence; this is the more
expected behavior, since the target was explicitly
authored into the asset.

(Internal change: 1960059)