[usdMaya] include non-registered variant set attributes on assemblies when collecting selections
In pipelines where variant sets have been registered, the reference
assembly would only consider variant set selections for those particular
variant sets, even if the assembly node had attributes on it for other
valid variant sets that had not been registered. Those selections for the
non-registered variant sets would then be ignored when opening and
composing the stage.

This change makes it so that all variant set selection attributes on the
assembly are considered in addition to all registered variant sets.

(Internal change: 1960309)