trace: specialized append-only event container
Implement an append-only array with an embedded doubly-linked list
node.  This simplifies the iterator implementation significantly and
also allows us to avoid separate allocations for the events array and
the list nodes.  Even though TraceEventContainer was careful to
reserve capacity so that its held vectors never reallocated, there was
no way to prove that property to the compiler so it emitted dead code
into every file containing trace scopes.

(Internal change: 1960407)