tf: emit static token destructors via TF_DEFINE_PUBLIC_TOKENS
TfStaticData<T> invokes T's destructor in the event of racing
initialization.  For public static tokens, this means that the
compiler would be required to emit a copy of token struct's destructor
into every translation unit that accesses a public token.  Static
token destructors are large and rarely invoked in practice.

(Internal change: 1960862)