Add optional validation for edits to SdfLayer
SdfLayer now has a validation mode that can be enabled
by setting the SDF_LAYER_VALIDATE_AUTHORING environment
variable to 1. This is primarily useful for diagnostics.

When enabled, SdfLayer will check to ensure that new specs
being created and and fields being edited are registered
with the layer's schema and will issue a coding error if
they are not. These checks could be extended in the future
to perform deeper validation of field value types, etc.

(Internal change: 1961014)