[usdAppUtils,usdviewq] add a UsdAppUtils library
This new UsdAppUtils library provides a place for functionality that
is common between applications that view and/or record images of USD
stages.

There are a few different categories of Python command-line arguments
including those for specifying a camera, color correction, refinement
complexity, and frame ranges. This also adds a utility function for getting
a UsdGeomCamera from a USD stage using just its name rather than a full
path. The stage is searched for a camera prim matching that name.

(Internal change: 1961347)