[Tests] Disable hud on testusdviewEmbreeRenderer, update baselines.
Having the hud as part of the image diffs was causing failures when system fonts were updated, which isn't ideal.  Disabling the hud makes this test more reliable, although it relies on usdview to generate appropriate error output when it can't find the Embree renderer (so that we don't accidentally produce GL output).

(Internal change: 1961728)