Declare Maya plugins as type "python" in plugInfo.json so that their shared library isn't loadable by Plug outside of Maya
If the plugins are type "library", then loading the .so may seem to work if the
prerequisite Maya libraries are available, but will ultimately fail in
possibly nondeterministic ways. Prevent the Plug library from ever loading
these Maya plugins by changing them to type "python" and removing the
LibraryPath. This pattern is used in other Maya plugins that need to register
types with UsdMaya's registries.

(Internal change: 1962297)