trace: reduce overhead of appending events
Many containers choose to make default-constructed instances
allocation-free both to allow noexcept default constructors and on the
basis that a lot of instances will remain empty.  Thus, they save time
and memory for a very common case and shift runtime overhead to the
first insertion.

However, TraceEventContainer is not a general purpose container.  We
construct them to collect trace events when we enable tracing.  Is it
reasonable to optimize for the case where we enable tracing but
collect no trace events?

TraceEventContainer now allocates its first node when constructed and
every TraceEventContainer has at least one node, even moved-from
containers.  This allows us to flip the alloc-then-construct pattern
around.  The event is emplaced unconditionally and afterward, if we
find that the node has become full, we allocate a new node in
anticipation of the next event.  In a tight loop, the carried
dependencies make alloc-construct vs construct-alloc seem similar but
this also allows TraceEventContainer to point directly at the next
available event location so that construction does not depend on
reading through _back->_end.

(Internal change: 1969505)
(Internal change: 1969707)