Added a new debug flags editor widget
This new debug flags editor shows you all the available debug flags
in the session without having to rely on a hardcoded list of prefixes.
It also shows the descriptions of what the debug flags do.

In a follow up change, we'll be removing the Debug menu that is now
no longer needed.

(Internal change: 1970078)