sdf: New SdfPath implementation in two parts.
Previously, SdfPath represented paths as a prefix tree (or trie) with each node
representing a path component.  SdfPath instances pointed to the tree node
representing the path.  This change makes it so we have two prefix trees.  One
for the "prim part" of an SdfPath (root nodes, prim nodes, prim variant
selection nodes) and one for the "property part" (property nodes, target nodes,
relational attribute nodes, mapper nodes, expression nodes).  SdfPath changes
to point to a prim-part node, and optionally also a property-part node if the
path is a property-like path.

We do this for a few reasons.  First, we observe that in our use, the number of
property paths can be very large in production scenes (100s of millions to a
billion) but the number of unique property names is typically only about 1% of
this number.  By creating a separate prefix tree we can reap this gain.
Second, we use SdfPath to index values in SdfLayers.  During attribute value
resolution, we form the path to the desired SdfAttributeSpec in each layer we
visit.  It's essential that forming this path be fast to make value resolution
fast.  By storing a separate prefix tree for properties, we can build an
effective per-thread cache of 'root' property elements, dramatically reducing
this cost.

To ensure SdfPath's size (8 bytes) remains the same, we refer to the prim and
property parts using 4-byte "handles".  These handles are provided by a new
thread-optimized pool allocator built specifically for this purpose.

This change also changes the behavior of SdfPath's lexicographical less-than
order in a minor way.  Previously, a prim path's prim children would sort
before its properties.  Now a prim path's properties sort before its prim
children.  For example, previously ['/a', '/a/a', '/a.a'] was in
lexicographical order, now it is ['/a', '/a.a', '/a/a'].  This detail was not
previously specified, and we don't expect this to cause issues.  The few tests
that failed due to this were simply recording the existing order, not relying
on it for any correctness purpose.

(Internal change: 1971003)
(Internal change: 1971040)
(Internal change: 1971238)
(Internal change: 1971354)
(Internal change: 1972108)