[usdMaya] fix for importing single transform xformOp prims that include shearing
This change re-implements the _pushUSDXformToMayaXform() helper function in
translatorXformable so that it accounts for any shear present in the
local transformation matrix of the USD prim being imported. It also
combines the time-sampled and non-time-sampled code paths into a single
loop through the time codes where in the non-time-sampled case, the vector
of time codes will contain only the default time.

(Internal change: 1971930)