[Hdx] Add HdxPickTask, for running picking queries.  Delete HdxIntersector.
The main goal of this change is to replace HdxIntersector (an opaque API that performed a pick operation) with a picking task that's more idiomatic for Hydra and can exist as a citizen of the render graph.  This fixes up some API headaches for us and paves the way, architecturally, for backend-independent picking.

The intent of this task was to preserve existing behavior.  HdxIntersector::Query() was necessarily broken into a few parts, to fit into task Sync() and Execute(), but the behavior should be the same.  Callers have been updated for the new API.

(Internal change: 1971956)