[Hd/HdSt/Hdx/Maya] Remove Multi-Collection Support from HdxRenderTask
This change is part of a series of changes in to fix the "collection paradox".

As part of the collection paradox fixes, ownership of rendertag is being removed from collections and added to tasks.
For the most part, tasks have a single collection, so moving ownership is relatively trivial.

Multi-Collection support was added to HdxRenderTask, so that Hydra could support the "Points" representation of
a prim.  This is previously a collection could only represent one Repr.  As we've moved to Repr selection, this usage
has been removed.

So the Maya front-end remains the final user of this multi-collection render task.  It does this, so that each "Shape" can
have a diffirent set of render tags.  To, for example, turn on Guides for a shape.

However, with ownership moving to the tasks.  This starts becoming more complex as the tasks would need to support
multiple render tag lists.

Therefore, we've decided to simply the problem, by restricting HdxRenderTask to one collection and set of render tags.

To keep the functionality in Maya, the Maya scene delegate now creates a HdxRenderTask per collection.  The
performance concern in this area is in the number of render passes.  Previously, the multi-collection render task would
create a pass per collection.  So with this change the net number of render passes remains the same.

Changes:
- Removed the definition of HdRprimCollectionVector, as this is no longer a core principle.
- Added extra debug information to HdChangeTracker::MarkTaskDirty(), so that which task missed is reported in the error message, making tracking down the error easier.
- Changed HdStDrawTarget's collection behaviour, so it no longer processes arrays of collections.  It was a coding error and additional collections beyond the 1st were ignored
  anyway
- HdxRenderTask switched to using a single collection.  As a result, render task now contains a single (optional) render pass.  (There are certain tasks that set an
  empty/default constructed collection, when the stage is disabled).
- The Maya Scene delegate method PxrMayaHdSceneDelegate::GetRenderTasks() has been updated to create a render task per collection.
    - The _renderTaskIdMap needed to be updated to support both the hash and collection name as a key.
    - The collection name needed to be sanitized (TfMakeValidIdentifier) to be supported in an SdfPath.
    - The new task name is formed by concatining in the collection name with the previous name of the render task token + hash.

(Internal change: 1972091)