[Hd/Hdx] Implement Render Tag Gather Phase
This change is part of a series of changes in to fix the "collection paradox".

The collection paradox is a data race within Sync.  The dirty list needs the render tags in order to perform filtering.
These tags were stored in a collection.  However, some tasks (draw targets and shadow tasks) obtained their collection from the corresponding Sprims.
These Sprims need to be synced before the collection could be obtained from them.
However, we would like to sync only those dirty Sprims *And* we would like to sync the scene delegate before syncing Sprims.

In order to resolve this cycle, this change moves ownership of the render tags from the collection to the task.
Therefore, the render tag can be obtained without needing the Sprim.

The render tag gather step happens straight after task sync.  As the list of tasks can differ from different calls to Sync All, the active set of render tags can grow over time.
The way it is supported is by gathering the render tags for the tasks every time.  If the active set is known to have changed (because the version was incremented), this
gathered set replaces the current active set.  If the active set has not changed, then the union of the gather and active set is created and if it differs from the active set,
to replaces the current active set and the active set is marked dirty, so that downstream caches will be invalidated.

Changes:
- Added a member _activeRenderTags to render index, that keeps track of the current set of render tags.
- A new method _GatherRenderTags() is used to update _activeRenderTags.
- The _activeRenderTags are now passed to HdDirtyList::GetDirtyRprims() instead of the set obtained from the collection.
- HdTask has a new GetRenderTags() API, with the default implementation returning an empty set of tags.
- Most tasks used a hard coded set of tags in the collection, so the Task have been updated to return those tags.
- Render/Render Setup is special because the Render Setup task owns the tags, but Render uses the tags in the Execute phase.

(Internal change: 1972104)