[Hdx] Add backend-independent picking to usdview.
This change adds a new task, HdxPickFromRenderBufferTask.  This has the same interface as HdxPickTask, but instead of generating an id render on the spot, it uses main-cam ID buffers passed in from the task graph.  The pick frustum is remapped onto the ID buffers to generate a sub-region of the ID buffers to iterate over.

HdxTaskController will generate an HdxPickFromRenderBufferTask on supported backends, as an alternative to HdxPickTask.

(Internal change: 1972300)
(Internal change: 1972324)