Fix UsdImaging crash with triply nested point instancer.
UsdImaging's PointInstancerAdapter was working with singly-
and doubly-nested UsdGeomPoinstInstancer prims, but would
crash when a third level of nesting was added.

The root cause was a mix-up between USD paths and cachePaths.
Specifically, cache paths for instancers may embed variant
selections to disambiguate uses of a USD point instancer in
the cache.  When looking up the backing USD prim, we must
strip off the variant selections -- but this step was missing
in some places.  This was relatively easy to spot after
adjusting variable names to distinguish paths used for
USD vs. cache paths.

Also, UpdateForTime() must check if the parent instancer prim
exists, similar to what UpdateForTimePrep() already does,
an issue revealed once the above was addressed.

Fixes #838

(Internal change: 1972877)