[usdImaging] add a 'usdrecord' command line tool for generating images from a USD file
This change includes two other main additions. The first is a
UsdAppUtilsFrameRecorder class that wraps a UsdImagingGLEngine and can be
used to output images to disk that are effectively the same as what you
would see in the viewer in usdview. The other is a rendererArgs.py module
that supports added command-line args to a script for specifying the Hydra
renderer plugin to use.

(Internal change: 1972902)
(Internal change: 1973128)