UsdImaging native instancing crash.
In this crash, we repopulate a native instancer adapter after some
scene edits.  Since this is a top-level Populate() call, the
instanceChainPath is </> and _GetPrimPathFromInstancerChain()
returns a path that does not reflect a real prim:

    instancePath: /__Master_2/obj_0
    parentProxyPath: /
    instanceChainPath: /obj_0

Work around the crash by checking if the returned prim is
valid before querying its modelDrawMode.

Fixes #847

(Internal change: 1972953)