[Hd, Hdx] Switch Hydra Tasks Execute Phase to use task-based Render Tags.
Prior to this change, tasks has two mechanism to specify render tags to the Execute Phase.
Firstly there was the render tags specified by the collection, which was used in the gathering of draw items.  Each render tag was bucketed into different command buffers.
Secondly, the task could pass a set of render tags to the Execute function, which would be used to select which render tag buckets to draw.

This change, removes the render tag bucketing and instead changes render pass to using a single command buffer.  Instead of using the collections render tags,
the render tags passed to Execute are used to do draw item filtering.

This simplifies the logic and provides a consistent set of render tags across sync and execute.  However, this will impacting render tag change performance as changing
render tags now causing command buffers to be rebuilt.  However, this isn't a high frequency operation and the new performance should be in line with Repr
(display mode switching) performance it was felt to be a good compromise.

Changes:
- Render Index's GetDrawItems() now takes render tags as a parameter and not longer returns a bucketed list of draw items, but instead just a single list of draw items.
- Render Index gather filter will now early out if a tag match is found.
- render tags is now carried separately through the draw item gather code, ignoring render tags in collections (which will soon be removed).
- Removed the various overloads of HdRenderPass::Execute(), as they all ended up just passing a render tag token vector to _Execute().  Now just a single entry point for Execute() exists.
- Most tasks now pass the result of it's own GetRenderTags() method to Execute().
- HdSt_RenderPass, now only contains a single command buffer.  Rebuilding if the render tags change
- Shadow task no longer uses a hard coded set of render tags for execute and instead uses those provided by GetRenderTags().

(Internal change: 1973001)