[Hd, UsdImaging, UsdImagingGL] Render delegates explicitly specify material adapter prim type tokens
We need UsdImagingDelegate to be able to create a custom material
adapter that would create shader resources for our new Maya plug-in
instead of HdSt ones. This was previously impossible because
UsdImagingDelegate selected one of the two hardcoded material adapter
prim type tokens based on the material binding purpose (full or
preview) returned by the render delegate.

We're instead suggesting a new virtual method that returns the prim
type token explicitly, thus allowing the render delegate to specify a
custom material adapter, transparently to UsdImagingDelegate.