Locally disable placement-new warnings in later versions of gcc.
Our use of boost::function in tf/pyFunction.h causes placement-new
warnings on newer versions of gcc and older versions of boost.

We can't easily get rid of boost::function yet because of
its pervasive use in the rest of the tree that depends on this
functionality in pxr.

(Internal change: 1973568)