[pxrUsdMayaGL] make shape adapters responsible for sanitizing rprim collection names
The names of the HdRprimCollections managed by Maya/Hydra shape adapters
now need to be sanitized since they are used as part of the SdfPath that
identifies render tasks. This change makes the shape adapters themselves
responsible for that so that the batch renderer does not need to sanitize
every collection name at every draw call.

(Internal change: 1973804)