[Maya] Remove references to render tags in collections from Maya Scene Delegate.
Part of the larger set of changes to cleanup usage of render tag, this change tackles the Maya scene delegate.
Hydra no longer uses the render tags as specified in the HdRprimCollection class as so it is desirable to remove this from the interface to prevent confusion.

However, the Maya Delegate is current using the HdRprimCollection to carry the render tag information.  For the  Maya delegate design it makes sense to pair
this information together, as it comes from the shape adapters and specified to the tasks together.

As both collections and render tags are used to filter the prim index, this change introduces a new data structure PxrMayaHydraPrimFilter which is used to pair this information.
Changes:
- Improved error reporting of PxrMayaHdSceneDelegate _GetValue() method.
- PxrMayaHdSceneDelegate _GetValue() was doing a double type check when extracting the value from the value cache.
- Renamed Batch Render's PopulateCustomCollection() method to PopulateCustomPrimFilter() as it now populates a prim filter instead of just a collection.
- Removed setting of render tags in the viewport rprim collections.
- Batch Render API's that took a collection or collection vector now take a prim filter or prim filter vector.
- Shape Adapter now has a method for obtaining the render tags.
- The Shape Adapter implementation now populate the base classes' _renderTag attribute rather than storing it in the collection.  There is no longer a need to mark the collection dirty.

(Internal change: 1974127)