tf: remove some function wrappers from python module processing
WalkModule is only used with member functions of Tf_ModuleProcessor so
we can replace std::function+bind with a pointer-to-member-function.
Doing so requires that we be consistent with signatures so
WrapForErrorHandlingCB and FixModuleAttrsCB have been changed to take
objects by const reference.  In WrapForErrorHandlingCB, this involved
changing some attr assignments to use setattr instead.

Most calls from Python into C++ go through InvokeWithErrorHandling so
we'd like to minimize the number of function wrappers that are
involved.  Rather than having a function, wrapped in std::bind,
wrapped in make_function, wrapped in raw_function, we pass a function
object that takes raw arguments to raw_function.

(Internal change: 1977518)