[usdImaging] Add support for inherited primvars to native instances.
Inherited primvars that are inherited through an instance boundary are automagically turned into instance-rate primvars on the hydra instancer for a given master.  For example, with the scene:
/World/A -> /_Master_1
/World/B -> /_Master_1
and displayColor defined as a constant primvar on both A and B, usdImaging will construct an instance-rate primvar on the /_Master_1 instancer composed of displayColor sampled at each instance path.

(Internal change: 1977568)