trace: avoid reading an invalid element after allocating a new node.
When change 1969505 switched the design to allocate in anticipation of
future events, it allowed the last node in the container to be empty.
This introduced a subtle bug in TraceEventContainer::back() because
the last element in the container could be in the previous node.

This went mostly unnoticed because very few callers actually read from
the reference returned by TraceEventList::EmplaceBack.  However,
TraceCollector::BeginEvent and EndEvent do so to return the timestamp
of their events.  The TRACE_* macros use timespan events but the
high-level scopes captured by performance tests typically use
Begin/EndEvent.  The bug manifested as performance metrics with
intermittent spikes to large values.

The solution here is to make TraceEventContainer::emplace_back()
return the reference itself (as in C++17) and eliminate back()
entirely.

(Internal change: 1978322)